+++ actix-web/CHANGES.md
@@ -9,7 +9,8 @@
 - Add `web::Payload::to_bytes[_limited]()` helper methods.
 - Add missing constructors on `HttpResponse` for several status codes.
 - Add `http::header::ContentLength` typed header.
-- Add Implement Default and serde::Deserialize for 'actix_web::Data'
+- Implement `Default` for `web::Data`.
+- Implement `serde::Deserialize` for `web::Data`.
 
 ### Changed